requirements.txt
Flask==2.5.2
scikit-learn==1.4.1.post1
matplotlib==3.7.1
Pillow==9.3.0
numpy==1.24.3
rasterio==1.4.10
dask==2023.8.1
detectree==0.1.1
